requirements.txt
django==3.1.0
djangorestframework==3.11.1
Pillow==7.1.2
celery==4.2.1
kombu==4.6.11
billiard==3.5.0.5
requests==2.23.0
mysqlclient==2.0.1
faiss-gpu==1.6.3